+++ Source/Podio .NET/Services/HookService.cs
@@ -19,10 +19,10 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="refType"></param>
         /// <param name="refId"></param>
-        /// <param name="externalURL">The url of endpoint.</param>
+        /// <param name="externalURL">The url of endpolong.</param>
         /// <param name="type">The type of events to listen to, see the area for options.</param>
         /// <returns></returns>
-        public async Task<int> CreateHook(string refType, int refId, string externalURL, string type)
+        public async Task<long> CreateHook(string refType, long refId, string externalURL, string type)
         {
             string url = string.Format("/hook/{0}/{1}/", refType, refId);
             dynamic requestData = new
@@ -31,7 +31,7 @@ namespace PodioAPI.Services
                 type = type
             };
             dynamic response = await _podio.Post<dynamic>(url, requestData);
-            return (int) response["hook_id"];
+            return (long) response["hook_id"];
         }
 
         /// <summary>
@@ -39,7 +39,7 @@ namespace PodioAPI.Services
         ///     <para>Podio API Reference: https://developers.podio.com/doc/hooks/delete-hook-215291 </para>
         /// </summary>
         /// <param name="hookId"></param>
-        public async Task<dynamic> DeleteHook(int hookId)
+        public async Task<dynamic> DeleteHook(long hookId)
         {
             string url = string.Format("/hook/{0}", hookId);
             return await _podio.Delete<dynamic>(url);
@@ -52,7 +52,7 @@ namespace PodioAPI.Services
         /// <param name="refType"></param>
         /// <param name="refId"></param>
         /// <returns></returns>
-        public async Task<List<Hook>> GetHooks(string refType, int refId)
+        public async Task<List<Hook>> GetHooks(string refType, long refId)
         {
             string url = string.Format("/hook/{0}/{1}/", refType, refId);
             return await _podio.Get<List<Hook>>(url);
@@ -60,12 +60,12 @@ namespace PodioAPI.Services
 
         /// <summary>
         ///     Request the hook to be validated. This will cause the hook to send a request to the URL with the parameter "type"
-        ///     set to "hook.verify" and "code" set to the verification code. The endpoint must then call the validate method with
+        ///     set to "hook.verify" and "code" set to the verification code. The endpolong must then call the validate method with
         ///     the code to complete the verification.
         ///     <para>Podio API Reference: https://developers.podio.com/doc/hooks/request-hook-verification-215232 </para>
         /// </summary>
         /// <param name="hookId"></param>
-        public async Task<dynamic> Verify(int hookId)
+        public async Task<dynamic> Verify(long hookId)
         {
             string url = string.Format("/hook/{0}/verify/request", hookId);
             return await _podio.Post<dynamic>(url);
@@ -77,7 +77,7 @@ namespace PodioAPI.Services
         ///     <para>Podio API Reference: https://developers.podio.com/doc/hooks/validate-hook-verification-215241 </para>
         /// </summary>
         /// <param name="hookId"></param>
-        public async Task<dynamic> ValidateHookVerification(int hookId, string code)
+        public async Task<dynamic> ValidateHookVerification(long hookId, string code)
         {
             string url = string.Format("/hook/{0}/verify/validate", hookId);
             dynamic requestData = new
