+++ Source/Podio .NET/Services/ImporterService.cs
@@ -20,24 +20,24 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="fileId"></param>
         /// <returns></returns>
-        public async Task<ImporterInfo> GetInfo(int fileId)
+        public async Task<ImporterInfo> GetInfo(long fileId)
         {
             string url = string.Format("/importer/{0}/info", fileId);
             return await _podio.Get<ImporterInfo>(url);
         }
 
         /// <summary>
-        ///     Imports the file into the given app. The mapping value for a field depends on the type of field. See the area for
+        ///     Imports the file longo the given app. The mapping value for a field depends on the type of field. See the area for
         ///     datails.
         ///     <para>Podio API Reference: https://developers.podio.com/doc/importer/import-app-items-212899 </para>
         /// </summary>
         /// <param name="fileId"></param>
-        /// <param name="appId">The id of the app the values should be imported into</param>
+        /// <param name="appId">The id of the app the values should be imported longo</param>
         /// <param name="mappings">The mappings between fields and columns</param>
         /// <param name="tagsColumnId">The id of the column to read tags from, if any</param>
         /// <param name="appItemIdColumnId">The id of the column to read the app item id from, if any</param>
         /// <returns></returns>
-        public async Task<int> ImportAppItems(int fileId, int appId, List<ImportMappingField> mappings, string tagsColumnId = null,
+        public async Task<long> ImportAppItems(long fileId, long appId, List<ImportMappingField> mappings, string tagsColumnId = null,
             string appItemIdColumnId = null)
         {
             string url = string.Format("/importer/{0}/item/app/{1}", fileId, appId);
@@ -49,18 +49,18 @@ namespace PodioAPI.Services
                 app_item_id_column_id = appItemIdColumnId
             };
             dynamic response = await _podio.Post<dynamic>(url, requestData);
-            return (int) response["batch_id"];
+            return (long) response["batch_id"];
         }
 
         /// <summary>
-        ///     Imports the given file into the space contacts on the given space.
+        ///     Imports the given file longo the space contacts on the given space.
         ///     <para>Podio API Reference: https://developers.podio.com/doc/importer/import-space-contacts-4261072 </para>
         /// </summary>
         /// <param name="fileId"></param>
         /// <param name="spaceId"></param>
         /// <param name="mappings">The mappings of the fields on the contact</param>
         /// <returns></returns>
-        public async Task<int> ImportSpaceContacts(int fileId, int spaceId, List<ImportMappingContactField> mappings)
+        public async Task<long> ImportSpaceContacts(long fileId, long spaceId, List<ImportMappingContactField> mappings)
         {
             string url = string.Format("/importer/{0}/contact/space/{1}", fileId, spaceId);
             dynamic requestData = new
@@ -68,18 +68,18 @@ namespace PodioAPI.Services
                 mappings = mappings
             };
             dynamic response = await _podio.Post<dynamic>(url, requestData);
-            return (int) response["batch_id"];
+            return (long) response["batch_id"];
         }
 
         /// <summary>
-        ///     Imports the file into the given app. The mapping value for a field depends on the type of field
+        ///     Imports the file longo the given app. The mapping value for a field depends on the type of field
         ///     <para>Podio API Reference: https://developers.podio.com/doc/importer/import-app-items-212899 </para>
         /// </summary>
         /// <param name="fileId"></param>
         /// <param name="row"></param>
         /// <param name="mappings"></param>
         /// <returns></returns>
-        public async Task<RowPreview> GetPreview(int fileId, string row, List<ImportMappingField> mappings)
+        public async Task<RowPreview> GetPreview(long fileId, string row, List<ImportMappingField> mappings)
         {
             string url = string.Format("/importer/{0}/preview/{1}", fileId, row);
             dynamic requestData = new
